+++ NEWS.md
@@ -1,5 +1,6 @@
-
-# flattabler (development version)
+# flattabler 1.1.0.9000
+* Change the implementation of [read_excel_sheet()], [read_excel_file()] and [read_excel_folder()] functions to use the readxl package.
+* Parameter [allSheets] to obtain all the sheets of all the Excel files in a folder: [read_excel_folder()].
 
 # flattabler 1.1.0
 * Functions to work directly with Excel files: [read_excel_sheet()], [read_excel_file()] and [read_excel_folder()].